Lower FRem as a call to CORINFO_HELP_{FLT,DBL}REM.
The backend will fail to emit functions containing FRem instructions
on platforms where fmod is not available at runtime; this currently
includes CoreCLR on Windows (by default, at least).

We may be able to defer this lowering by changing the target library
options for the backend, but this likely requires a change to MCJIT.

Fixes #19.