Implement the simple part of methodNeedsToKeepAliveGenericsContext and related methods
Implement the simple part of methodNeedsToKeepAliveGenericsContext, where we don't actually need to keep anything alive.

Implement thisObj and related parameter fetchers.

Implement derefAddress  which is used to tunnel into opaque EE data. Lots of casting required here.

Update the vtable type we use for objects to indicate it points at an array of pointer-sized things, and fix a couple of other char *'s to be intptr_t *'s for similar reasons.

Factor getFunctionType so we can call it on a signature, use this to cast vtable fetches to the right types.

We now handle a few more methods in the initial test suite, and have updated fail reasons on a bunch more.