Update RunTests to Report Failure for Diffs or Failures
We should fail if there are any diffs or any failures in RunTests.
This change updates CheckDiffs to return the number of diffs during a
run. This change also updates the path to the test log since it has
changed. RunTests will return False if there are any diffs or failures
and will return True if there are no diffs and no failures.

This change also updates CheckDiffs to use Write-Host to print messages
to output so that we do not pollute the return value with side effects.