Implement genNullCheck
Generate an explicit compare and conditional branch to a block that throws
a NullReferenceException (though using DivideByZeroException as a stand-in
pending a new helper method provided by the EE).

The block properties are updated so that the 2nd pass worklist processing
will see the block with the check as ending and move on to process the
successors, which will include the new block that was split out to hold
the non-exception path.

There's a bit of subtlety in that blocks are annotated with Begin/End MSIL
Offsets as though each MSIL instruction falls neatly into some block, but
here we have instructions which have an exception check in one block, the
code to raise the exception if the check fails in a second block, and the
code that completes the operation in a third block.  The convention I've
chosen is to set the Begin and End offsets of the throw block (which need
to be equal to each other to ensure that subsequent processing doesn't try
to parse MSIL and insert more instructions in to this block) both equal
to the offset at which the exception-bearing instruction begins; to
set the Begin of the non-exception successor to the offset at the end of
the instruction (to avoid having the worklist processing re-parse the
instruction and re-generate it in that block), and to set the End of the
predecessor to the offset at which the exception-bearing instruction ends
(because this makes the code in genNullCheck slightly more compact).  If
we need to we can come back and add some more state to the block
annotations to make this more explicit (or add accessors that check for
this pattern), but as it stands now there's no code that would need to
consume them.