Remove volatile check for call
We unnecessarily check volatility in call helper where we always emit call instead of intrinsic for certain cases.
I removed it and added comments below when we are about to use intrinsic.
Even for the case of call, we still need to add barrier for volatile operation.
I implement memory barrier using the existing llvm fence construct which uses MFence for sequential consistency.