Defer Check for DiffTool to CheckDiff
If we are not using the difftool to do any diffs, we should not be
required to have the diff tool on the path. Therefore, we should defer
checking if sgdm.exe is available to the CheckDiff function, and only
check for it if the user has specified that they want to use the diff
tool.