Clear NuGet Cache and fix diff tool checking problem.
NuGet uses a cache for downloaded package so that a later NuGet request
does not have to download again and uses a previously downloaded package.
However sometimes on some machine it caused some unpredicted behavior
that led to different versions of CoreCLR/CoreFX packages. Finally the
difference is reflected in LLVM IR diffs we are catching. To exclude any
of the unpredictable factors, an operation is performed to clear NuGet
cache in the beginning of a session. Also fix a bug due to delaying
diff tool check to CheckDiff.