Document GenStack class and improve interface and implementation
This change documents the GenStack class. In addition we modify
the base class, ReaderStack, to specify that its state is
represented by a std::vector. We then also change the iterator
interface so that it used standard iterators. We provide
iterators that iterate from the bottom to the top of the stack
as that seems to be what is needed by our current algorithms.
In addition the depth() method was renamed to size() to be consistent
with STL conventions.

The simple methods that are not expected to require client
customization are implemented in ReaderStack (e.g. begin()).
Those that might need client-specific diagnostics or
precondition checks (e.g. push and pop) are implemented
in GenStack.

Also remove unused ReaderStackIterator type

Closes #121.