Add ability to format all files in ccformat
We want to be able to run ccformat in the lab as a checkin barrier. #232
also requests that ccformat run over the full LLILC code, rather than
just the diffs. This change adds a --formatall switch which triggers a
full clang-format run over all of our sources and prints out which files
need to be reformatted. Because we are not looking at just the files that
have changed, we run clang-format directly, so we add the --clang-format
switch for the user to specify where clang-format is if not on the path.
We skip the .git and include\clr directories since the .git directory is
not source and the include\clr directory has clr headers that we do not
want to change. We also only run clang-format on c, cpp, and h files, so
we are not checking CMakeFiles.txt files. If there are diffs, we
instruct the user to run with --fix.

This change also updates our regular format run to print out which files
need to be reformatted, rather than printing the diff between the fixed
and original file, since the diff is pretty unreadable. We also fail if
there are formatting changes necessary and tell the user to run with
--fix.