Add diff checking to formatall
We want the diff for formatall like we have with the regular format checking.
This change copies logic from clang-format-diff.py that compares the
before and after.

Add --noindex option to ccformat

Currently in the lab, diffing based on branches does not work. This
change adds the --noindex option, which allows the user to diff two
repositories currently on their machine, rather than comparing with a
remote branch. This will primarily be used in the lab where we will
clone the most recent prtest branch as well as the master dotnet branch
of llilc.

This change also exposes the returncode of the script so if we fail for
any reason the lab job will fail with the exit code.