Allow user control whether to update test assets
Give user a control whether to update test assets every time
when the script starts. In some cases, user does not want to
automatically update the test assets every time. By default,
the test assets update is on every time. To throttle that,
call the script with swith -NoTestUpdate, no value is needed
to pass into this parameter. It will be $True for a switch.