Reformat sources to be LLVM convention compliant
Several files were missed when we ran ccformat on our source base. This
change reformats these files so they are LLVM convention compliant.

Add .clang-format to get rid of the style errors. Add clang-format on
and off comments to LLILCPal.h and readerir.h to stop trying to format
certain parts of the code.

Factor out opcodes array to bytecodetowvmcode.def so clang-format
doesn't format it.