Work on the entryLabel NYI failures
`entryLabel ()` was used to enumerate the optimistic tail recursion branches. Since we don't do this optimization we don't need to ask for the entry label. I added a flag `HasOptimisticTailRecursionTransform` to `ReaderBase` that the client can set if such branches are added, and simplified `fgFixRecursiveEdges` to fail only if it sees that flag and `HasLocAlloc`.

Net result is that the two cases that failed with this NYI -- Path.NormalizePath and PathHelper.GetFullPathName -- now fail handling mul.ovf, and those mul's feed locallocs. So no net gain on methods passing but we gain bit more clarity on what it will take to get them handled properly.