Add missing case to fgSplitBlock
Handle the case that a block has no terminator but does have one or more
instructions after the split point.  This can happen when inserting a null
check during the post-processing of a newobj (which generates the call and
then rewinds the insert point to do the post-processing which inserts the
null check) in a block that ends with a return (and thus doesn't get a
terminator in the first pass).