Make ccformat --formatall compatible with Python 3
Format all was incompatible with Python 3 because proc.communicate()
returns bytes. This change fixes formatall so that we are compatible
with python 3. This change also makes the join call in formatall
compatible with Python 3 and prints a new line at the end of diffs for
diff reability.

This change also fixes a formatting error that was checked in in
reader.h