Implement finally continuation selection for leave
For each finally, generate a continuation selector variable, and end the
finally with a switch that jumps to a continuation based on the selector.

For each leave instruction, set the selector variables appropriately for
each finally handler being exited before jumping to the innermost finally.

These explicit sequences match what Clang generates for gotos across
cleanups and what the WinEHPrepare funclet extractor will expect to see.

The 'default' case for each switch targets a block with just an
Unreachable in it (again following Clang's example).

The instructions to set the selector variables are generated and inserted
in the IR during first-pass flow-graph building.  I've added a map to the
reader objects tracking where selector stores have been inserted at which
MSIL offsets, updated fgNodeGetEndInsertIRNode to consult the map, and
updated a few places that were manually grabbing a block's terminator to
instead call fgNodeGetEndInsertIRNode, so that 2nd pass IR insertion can
use the correct insertion point when it adds IR to these blocks.
The switch instruction itself (and selector load feeding it) is generated
when processing the first leave across a particular finally, and inserted
when processing the first endfinally for that finally.  Since a finally
region can have an arbitrary number of endfinally instructions, endfinally
becomes a branch that targets a block that holds the switch, so that all
endfinally instructions for a given finally can branch to the shared
switch.

Resolves #260