Ensure generic context is kept alive.
This gets us the next part of #40.

We use `llvm.frameescape` to indicate that the context-bearing local's address has escaped. This intrinsic requires that the function have a frame poniter, so disable FPO for these methods.

 We're still missing the ability to report the context back to the EE via the GC info. Opened Issue #336 for this.

Also fixed a bug in the static field address computation where we weren't capturing the updated helper call result.