Implement conditionalDerefAddress.
This method checks the low bit of address, and if set, clears it and dereferences. Otherwise it just returns the unmodified address.

This generates plausible looking code but the method that contains it also has a convertHandle, so the codegen hasn't been tested at runtime yet.

As part of this I refactored the existing conditional expansion code somewhat. No diffs in IR other than changing the failure reason for one method.

Closes #356.