Remove checked in CoreCLR headers
Remove CLR headers and add dependency on a prebuild CoreCLR. Most of the headers in the
include\clr subdirectory are removed and now are picked up from a prebuilt coreclr so that
keeping in sync is clearer.

Remaining headers are used to support packing in the PAL.  Follow on
work will remove these dependencies.