Fix GenIR::castOp for {isinst,castclass} [valuetype].
GenIR::castOp incorrectly assumed that the target type of a cast
operation (i.e. castclass or isinst) must be a reference type.
According to ECMA-335, it is legal for the target type to be a value
type: under such a circumstance, the target type is interpreted to be
the boxed version of that value type.

I had hoped to simply use ICorStaticInfo::getTypeForBox to calculate
the boxed type (as in makeBoxDstOperand), but as it turns out, this
confusingly-named method only handles normalization of Nullable<T>.
If the argument to this method is not an instantiation of Nullable<T>,
it simply hands the argument back unchanged. As a result, I had to
write a helper to calculate the boxed type.