Fix two issues found when running Roslyn.
`loadAtAddress` of a struct type was always passing true for `IsField` when it is also used for array accesses. This caused failures for arrays of structs. Fix is to look at the ResolvedToken to determine whether the access is a field or not.

`conditionalDerefAddress` had a bug introduced when I refactored it. Since the method that hits this fails later in the reader the IR validation error hasn't yet surfaced. But Roslyn has methods where this bug causes validation errors.