Abstract the signature of the method being compiled.
Add a type that encapsulates the method signature and other relevant
details about the method being compiled. This lets us bring some logic
down into ReaderBase from GenIR and eliminates some code that would
otherwise need to be kept in sync between ReaderBase and its subclasses.

This is another step on the road to value type return values and parameters.