Add an ABI abstraction for computing function signatures.
LLVM's support for calling conventions does not handle aggregate
parameters or return values in a way that is consistent with the
rules for a particular calling convention. Instead of applying
whatever rules are particular to a given calling convention, it
simply recursively decomposes aggregates into a linear sequence
of arguments of non-aggregate type. It is the frontend's
responsibility, then, to emit calls such that the backend will
lower them appropriately for the target calling convention.

Practically speaking, this means that the frontend must do the
work to determine how each argument to a function--aggregate or
otherwise--should be passed (e.g. directly/indirectly, as an
expanded series of arguments, etc.) when generating calls and
function symbols. To this end, this change adds abstractions
to encapsulate ABI information and compute argument and result
passing information for a function signature (i.e. a calling
convention, result type, and argument types). These abstractions
are then used to emit function calls, home method arguments, and
return method results in a way that is compatible with the
backend.