More fixes from Roslyn.
Delete the LLVM EE when we're done jitting a method. This gets back most of the memory we allocate and brings our leaking down to a much more tolerable level.

Type the result of ldvirtfn as native int so that subsequent processing handles it the same as other function addresses.

Trim type array in the varargs constructor call path so the LLVM function type builder is not confused.

Add extra paranoia around ldlen and strlen since we see cases where the referent type is not an array or string. For now bail out if we hit one of these.

if we see an endfinally and can't locate the enclosing EH region, bail out with an NYI rather than AVing in the jit.