Statepoints: Don't report base-pointers explicitly.
For interior pointers, the CLR Runtime doesn't require the base-pointer
to be additionally reported explicitly. Therefore, this change turns off
base-pointer tracking and reporting at a statepoint.

This change also solves an issue with alloca sequence, where we make a
unmanaged to managed pointer cast.

For example, for ldloca, we generate

  %loc1 = alloca i8
  %1 = addrspacecast i8* %loc1 to i8 addrspace(1)*

With Base-pointer tracking the set of live poionters after these instructions
would be {%1, %loc} which is clearly wrong because %loc1 is an
unmanaged pointer i8*.

Without base-pointer tracking, the live set is correctly computed as {%loc1}