Use normal lowering for volatile stores.
There's a JitConfig setting JitLockWrite that should force the jit to emit lock prefixed stores for volatiles, but RyuJit doesn't look at this. So we'll also ignore it for now. Issue #278 will remain open to track implementing the option, if we decide to do so.

Instead just allow volatile stores to lower normally (with IsVolatile true in LLVM IR).