Fix bug in GenIR::storeElem.
When storing an element of struct type we call rdrCallWriteBarrierHelper.
We were passing array address instead of element address as the first argument.
This resulted in bad codegen when compiling Roslyn.

Change the return type of genArrayElemAddress from Value * to IRNode * and
eliminate a few casts.

Remove some unnecessary semicolons and fix a comment typo.