Call JIT_GCPoll() runtime helper from gc.safepoint_poll()
gc.safepoint_poll() helper required by Safepoint placement pass was an
empty function. This change fixes this to call the correct runtime helper
to perform the CLR GC Poll.

The current strategy is to use an unconditional call to the GCPoll helper.
Inlining the GCPoll helper is future work.

Issue #32