Precise GC: Bail early on managed aggregates
Precise GC using statepoints cannot handle aggregates that contain
managed pointers yet. (Issue #33)

So, this change adds a check to the reader-postpass, which scans the
generated IR if it contains values of managed-aggregate type.
If so, the compilation fails gracefully by throwing a NYI exception.

This check in ReaderIr.cpp is based on the assertion checks in
RewriteStatepointsForGC.cpp.

[Issue #32]