Add Python llilc_runtest script
This script runs CoreCLR test with specified LLILC JIT and pre-built
CoreCLR runtime. If verbose level is specified, either summary or
verbose, a result will be created in default location or a place
that is specified. In verbose case, the result will be normalized
with filter.

It is required to run this script from tests directory in a CoreCLR
repository.

To exclude undesired test cases, please edit exclusion file.

usage: llilc_runtest.py [-h] [-a {x86,x64}] [-b {debug,release}]
                        [-d {summary,verbose}] [-r RESULT_PATH] -j JIT_PATH -c
                        CORECLR_RUNTIME_PATH

optional arguments:
  -h, --help            show this help message and exit
  -a {x86,x64}, --arch {x86,x64}
                        the target architure
  -b {debug,release}, --build {debug,release}
                        release or debug build of CoreCLR run-time used
  -d {summary,verbose}, --dump-level {summary,verbose}
                        the dump level: summary, or verbose
  -r RESULT_PATH, --result-path RESULT_PATH
                        the path to runtest result output directory
  -j JIT_PATH, --jit-path JIT_PATH
                        full path to jit .dll
  -c CORECLR_RUNTIME_PATH, --coreclr-runtime-path CORECLR_RUNTIME_PATH
                        full path to CoreCLR run-time binary directory

This script assumes developer handles the work environment by themselves
and CoreCLR test has been built.

llilc_checkpass will follow.

This script is LLILC wrapper of underlying CoreCLR runtest which is
now still Windows only and about to get cross-platform. When that happens,
the implementation of this script will be updated accordingly. But the
command line interface is intended to remain the same.