Support indirect unmanaged calls.
Unmanaged calls require some extra machinery to setup the necessary data
structures and inform the runtime that control is entering unmanaged code.

This code is comprised of two pieces:
- Unmanaged call frame setup, which occurs in the method prolog, and
- GC mode transition, which occurs at an unmanaged call site.

Unmanaged call frame setup is relatively simple and can be accomplished
entirely in LLVM IR.

The GC transition, however, is more complicated due to the constraints
that must be satisfied for the GC to operate correctly when unmanaged code
is running. In particular, managed code must not run if the GC has been
informed that unmanaged code is running. The approach taken by other
managed code generators (e.g. RyuJIT) in order to meet this constraint is
to inform the GC of the transitions to and from unmanaged code immediately
before and after the machine call instruction, respectively. The same approach
is taken here through the use of the @llvm.experimental.gc_transition intrinic,
which informs the backend to surround the call with the appropriate transition
code. Any code that is not part of the transition itself is represented
directly in IR.