Reposition first-pass point blocks
Point blocks created in the first pass might not yet have a spot for their
point available to insert them into the block list.  This change:
 1. adds tracking of whether the first pass is completed or not,
 2. initializes the NodeOffsetListArray earlier so it will be available
    when point blocks are craeted in readerPrePass,
 2. during the first pass (and pre-pass), updates point block creation to call
    fgAddNodeMSILOffset to ensure that an appropriate spot will be created
    for it in fgReplaceBranchTargets, and
 3. in replaceFlowGraphNodeUses (which gets called from
    fgReplaceBranchTargets), moves the point node to the appropriate spot.

This change also fixes some blocks that are currently out-of-order:
 1. Create endfinally blocks and unreachable continuation blocks as point
    blocks (so the processing above will pick them up)
 2. Add a PreviousNode parameter to makeFlowGraphNode, use it to keep the
    tail block next to the ret block when we process ret.

This gets first-pass created point blocks in correct MSIL offset order
before we process them in the second pass.  The checking in msilToIR is
updated to assert that the blocks are already in sorted order.

Also fix some stale comments around fgReplaceBranchTargets.