Add initial Bring up Milestones markdown.
This adds a place for us to keep milestones that will take a PR.
These milestones will be linked to out of the wiki for reference. For
additions open a github issue to start a conversation then PR an edit
that has the new milestone.  When milestones are acheived strike through the txt.