Update individual tests exclusion mechanism in LLILC side
CoreCLR PR dotnet/coreclr#802 implemented an exclusion
mechanism for individual test cases which works generally for different JITs.
This is a corresponding test case exclusion mechanism update in LLILC side
to use that. Individual tests to exclude are specified in ExcluseList ItemGroup
in exclusion.targets file. With this change, we no longer hard delete built
test cases which should reduce a lot of issues and inconvinience for developers.
exclusion file removed. Both Powershell and Python script are updated.