Disable fast ISel.
The current implementation of the GC statepoint design is incompatible
with fast ISel, which may visit a call to @llvm.experimental.gc_result
before the corresponding call to @llvm.experimental.gc_statepoint.
Unfortunately, disabling fast ISel turns out not to be as simple as
setting TargetOptions.EnableFastISel to false; instead, the command-line
option must be set to 'false'.