Infer the type of 'this' in indirect calls.
Indirect call signatures may not specify the type of the 'this' parameter
of the target method. To handle this case, infer the type of the 'this'
parameter from the type of the 'this' argument. This is consistent with
the behavior prescribed by ECMA-335, which states that the type of the
'this' argument should be assumed to be correct in this case (see section
III.3.20).