Implement EndFault stub support
`endfinally` and `endfault` are the same MSIL instruction (distinguished by
whether the enclosing region's type is `fault` or `finally`).  Accomodate
both by changing the `getInnermostFinallyRegion` helper to also find
`fault`s (and rename it to `getInnermostFaultOrFinallyRegion`), adding
`fgMakeEndFault` parallel to `fgMakeEndFinally`, and invoking one or the
other according to the discovered region type.  Also get rid of the
superfluous `fgMakeEndFinallyHelper` routine.

`fgMakeEndFault` just generates an UnreachableInst for now (since handlers
are stubbed out).

Closes #393