Add Issue ItemMetaData to ItemGroup ExcludeList in exclusion.targets
Add Issue ItemMetaData to ItemGroup ExcludeList in excusion.targets
for each test case so that when a known issue is fixed we know which
tests to re-enable. Some of the issue numbers are relatively general
and more like a placeholder. As investigation goes on and we become
clearer about the causes, more specific issue number will replace the
placeholder.