Parse the string value of environment variable COMPLUS_AltJit.
The string value of COMPLUS_AltJit specifies which methods should
be fed to the alternate JIT.

The string may contain one or more patterns.  For example, the
string "*" contains one pattern, specifying that all methods
should be passed to the alternate JIT.

Similarly, the string "calc  System.Array:Sort  over(2)" contains
3 patterns.  It passes to the alternate JIT, any method called
"calc", irrespective of its namespace or class or number of
arguments; the method called "System.Array:Sort"; or any method
called "over", in any namespace, so long as it is called
with 2 arguments.

The patterns recognized are as follows:

  *   => all methods.
  M   => method M.
  C:M => method M in class C.

  C can be a simple class name, or a namespace.classname.

  M can be followed by the pattern (number) to specify its
  number of arguments.  For example, .ctor(0) or Adjust(2)