Update binaryOpType to handle more cases.
`binaryOpType` now roughly follows Ecma-335 table III.2 with some exceptions for internal use and for patterns seen in IL stubs. We are a bit more tolerant than the spec because we can't distinguish native int from int64 (or int32 on a 32 bit platform).

No diffs in the existing test cases.

Closes #558.