Disallow first-class aggregates in LLVM IR we are generating. They were causing struct type mismatches due to generic sharing and also had potential optimization issues.
No values of StructType are allowed on the operand stack.
Instead pointers to structs are pushed to the stack and
recorded in StructPointers set.

When a struct needs to be loaded we make a copy of the struct
and return a pointer to the copy.

addressOfValue doesn't need to copy the struct. Instead a pointer
to the struct is copied.

Struct type mismatches on the stack and at PHI joins
are handled by casting the pointers representing the structs.