Improve llilc_checkpass failure detection.
llilc_checkpass was previously checking for failures by hunting for lines
in a diff that began with '-'. Unfortunately, this is not accurate: most
importantly, it fails to account for the fact that methods may simply no
longer be submitted for compilation (generally due to changes to the
libraries or tests).

This change switches to a set-based implementation in order to accurately
detect methods meeting any of the following criteria:
- Compiled cleanly in the baseline and are now failing to compile
- Failed to compile in the baseline and are now compiling cleanly
- Processed in the baseline but not processed now
- Not processed in the baseline and processed now

A run is only a failure if there are methods that used to compile cleanly
that now fail to compile. A count for each category is reported in the output.